requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
import-export==0.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tablib==3.4.0
xlrd==2.0.1
xlwt==1.3.0
